+++ docs/notebooks/failure_ego.pct.py
@@ -22,7 +22,7 @@ import trieste
 
 def masked_branin(x):
     mask_nan = np.sqrt((x[:, 0] - 0.5) ** 2 + (x[:, 1] - .4) ** 2) < 0.3
-    y = np.array(trieste.utils.objectives.branin(x))
+    y = np.array(trieste.objectives.single_objectives.branin(x))
     y[mask_nan] = np.nan
     return tf.convert_to_tensor(y.reshape(-1, 1), x.dtype)
