+++ docs/notebooks/multi_objective_ehvi.pct.py
@@ -17,7 +17,7 @@ from trieste.data import Dataset
 from trieste.models import create_model
 from trieste.models.model_interfaces import ModelStack
 from trieste.space import Box
-from trieste.utils.multi_objectives import VLMOP2
+from trieste.objectives.multi_objectives import VLMOP2
 from trieste.utils.pareto import Pareto, get_reference_point
 from trieste.acquisition.rule import EfficientGlobalOptimization
 
@@ -33,7 +33,7 @@ tf.random.set_seed(1793)
 
 # %%
 vlmop2 = VLMOP2().objective()
-observer = trieste.utils.objectives.mk_observer(vlmop2)
+observer = trieste.objectives.observer.mk_observer(vlmop2)
 
 # %%
 mins = [-2, -2]
