+++ docs/notebooks/recovering_from_errors.pct.py
@@ -17,7 +17,7 @@ random.seed(3)
 
 # %%
 import trieste
-from trieste.utils.objectives import branin
+from trieste.objectives.single_objectives import branin
 
 
 class FaultyBranin:
