+++ tests/unit/acquisition/test_function.py
@@ -68,7 +68,7 @@ from trieste.models import ProbabilisticModel
 from trieste.space import Box
 from trieste.type import TensorType
 from trieste.utils import DEFAULTS
-from trieste.utils.objectives import BRANIN_MINIMUM, branin
+from trieste.objectives.single_objectives import BRANIN_MINIMUM, branin
 from trieste.utils.pareto import Pareto, get_reference_point
