+++ tests/unit/acquisition/test_sampler.py
@@ -32,7 +32,7 @@ from trieste.acquisition.sampler import (
 )
 from trieste.data import Dataset
 from trieste.space import Box
-from trieste.utils.objectives import branin
+from trieste.objectives.single_objectives import branin
 
 
 @pytest.mark.parametrize("sample_size", [0, -2])
