+++ tests/unit/objectives/test_multi_objectives.py
@@ -1,4 +1,4 @@
-# Copyright 2020 The Trieste Contributors
+# Copyright 2021 The Trieste Contributors
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -11,13 +11,20 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
 import numpy.testing as npt
 import pytest
 import tensorflow as tf
 
 from tests.util.misc import TF_DEBUGGING_ERROR_TYPES
 from trieste.type import TensorType
-from trieste.utils.multi_objectives import DTLZ1, DTLZ2, VLMOP2, MultiObjectiveTestProblem, vlmop2
+from trieste.objectives.multi_objectives import (
+    DTLZ1,
+    DTLZ2,
+    VLMOP2,
+    MultiObjectiveTestProblem,
+    vlmop2,
+)
 
 
 @pytest.mark.parametrize(
