+++ tests/unit/objectives/test_single_objectives.py
@@ -1,4 +1,4 @@
-# Copyright 2020 The Trieste Contributors
+# Copyright 2021 The Trieste Contributors
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -11,6 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
 from typing import Callable
 
 import numpy.testing as npt
@@ -19,7 +20,7 @@ import tensorflow as tf
 
 from trieste.space import Box
 from trieste.type import TensorType
-from trieste.utils.objectives import (
+from trieste.objectives.single_objectives import (
     ACKLEY_5_MINIMIZER,
     ACKLEY_5_MINIMUM,
     ACKLEY_5_SEARCH_SPACE,
@@ -51,7 +52,6 @@ from trieste.utils.objectives import (
     hartmann_3,
     hartmann_6,
     logarithmic_goldstein_price,
-    mk_observer,
     rosenbrock_4,
     scaled_branin,
     shekel_4,
@@ -107,26 +107,3 @@ def test_no_function_values_are_less_than_global_minimum(
 ) -> None:
     samples = space.sample(1000 * len(space.lower))
     npt.assert_array_less(tf.squeeze(minimum) - 1e-6, objective(samples))
-
-
-def test_mk_observer() -> None:
-    def foo(x: tf.Tensor) -> tf.Tensor:
-        return x + 1
-
-    x_ = tf.constant([[3.0]])
-    ys = mk_observer(foo, "bar")(x_)
-
-    assert ys.keys() == {"bar"}
-    npt.assert_array_equal(ys["bar"].query_points, x_)
-    npt.assert_array_equal(ys["bar"].observations, x_ + 1)
-
-
-def test_mk_observer_unlabelled() -> None:
-    def foo(x: tf.Tensor) -> tf.Tensor:
-        return x + 1
-
-    x_ = tf.constant([[3.0]])
-    ys = mk_observer(foo)(x_)
-
-    npt.assert_array_equal(ys.query_points, x_)
-    npt.assert_array_equal(ys.observations, x_ + 1)
