+++ trieste/utils/__init__.py
@@ -16,7 +16,7 @@ from typing import Callable, Mapping, TypeVar
 
 from typing_extensions import Final
 
-from . import objectives, pareto
+from . import pareto
 from .misc import DEFAULTS, Err, Ok, Result, T_co, jit, shapes_equal, to_numpy
 
 K = TypeVar("K")
